% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maddala.Cox.Snell.R
\name{Maddala.Cox.Snell}
\alias{Maddala.Cox.Snell}
\title{Maddala.Cox.Snell}
\usage{
Maddala.Cox.Snell(fit)
}
\arguments{
\item{fit}{Your model. Support logistic regression and Cox regression.}
}
\value{
Maddala Cox Snell.



#'@details The outcome variables in the model must be represented using 0 and 1. Among them,
          1 represents the occurrence of the event.
}
\description{
Maddala Cox Snell in the computational model.
}
\references{
Riley RD, Ensor J, Snell KIE, Harrell FE, Martin GP, Reitsma JB, et al. Calculating the sample size required for
           developing a clinical prediction model. BMJ (Clinical research ed). 2020
}
