% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prematurity.R
\docType{data}
\name{prematurity}
\alias{prematurity}
\title{A data on indicators for premature newborns.}
\format{
An object of class \code{data.frame} with 189 rows and 11 columns.
}
\usage{
data(prematurity)
}
\description{
A data on indicators for premature newborns.
}
\examples{
data(prematurity)

}
\keyword{datasets}
